requirements.txt
customtkinter==5.2.2
pillow==10.4.0
playsound3==3.2.2
pyautogui==0.9.54
